# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Bizosto ERP API
  version: 1.0.0
  description: >-
    Multi-tenant API contract generated from App Router route handlers.
servers:
  - url: /
security:
  - BearerAuth: []
...
paths:
...
  /api/files/{id}/restore:
    post:
      summary: POST /api/files/{id}/restore
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/files/share:
    post:
      summary: POST /api/files/share
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/integrations/quickbooks/callback:
    get:
      summary: GET /api/integrations/quickbooks/callback
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/sales_manager/deals/update:
    post:
      summary: POST /api/sales_manager/deals/update
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
  /api/sales/deals/close:
    post:
      summary: POST /api/sales/deals/close
      tags:
        - Core
      responses:
        '200':
          description: Request completed successfully.
        '401':
          description: Unauthorized or invalid session.
        '403':
          description: Forbidden for current tenant or role.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT